@@ -9,20 +9,29 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^10.0.0",
-    "expo": "^35.0.0",
-    "expo-font": "^7.0.0",
-    "native-base": "^2.13.8",
-    "react": "16.8.3",
+    "expo": "^36.0.0",
+    "expo-barcode-scanner": "~8.0.0",
+    "expo-font": "~8.0.0",
+    "expo-keep-awake": "~8.0.0",
+    "expo-location": "^8.0.0",
+    "expo-permissions": "^8.0.0",
+    "expo-secure-store": "~8.0.0",
+    "moment": "^2.24.0",
+    "react": "16.9.0",
     "react-devtools": "^4.2.0",
-    "react-dom": "16.8.3",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
+    "react-dom": "16.9.0",
+    "react-moment": "^0.9.7",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.0.tar.gz",
     "react-native-elements": "^1.2.7",
-    "react-native-gesture-handler": "~1.3.0",
+    "react-native-gesture-handler": "~1.5.0",
+    "react-native-gifted-chat": "^0.13.0",
     "react-native-modal": "^11.5.3",
-    "react-native-reanimated": "~1.2.0",
-    "react-native-screens": "~1.0.0-alpha.23",
+    "react-native-picker-select": "^6.3.3",
+    "react-native-reanimated": "~1.4.0",
+    "react-native-screens": "2.0.0-alpha.12",
     "react-native-web": "^0.11.7",
     "react-navigation": "^4.0.10",
+    "react-navigation-header-buttons": "^3.0.4",
     "react-navigation-stack": "^1.10.3",
     "react-navigation-tabs": "^2.5.6"
   },
