@@ -1,7 +1,9 @@
 import React, { Component } from 'react';
-import { View, StyleSheet, Text, ScrollView, TextInput, TouchableOpacity, Alert } from 'react-native';
-import color from '../constants/Colors';
+import { View, StyleSheet, Text, ScrollView, ActivityIndicator, TextInput, TouchableOpacity, Alert } from 'react-native';
+import Color from '../constants/Colors';
 import str from '../constants/Strings';
+import * as SecureStore from 'expo-secure-store';
+
 import getEnvVars from '../environment';
 const { apiURL } = getEnvVars();
 
@@ -10,161 +12,163 @@ class AddPostScreen extends Component {
     constructor(props) {
         super(props);
         this.state = {
-            loading: true,
+            isLoading: false,
             title: '',
             text: '',
-            titleIsValid: false,
-            newPost: {
-                title: '',
-                text: '',
-            }
+            userId: ''
         };
     }
 
-    componentWillMount() {
+    UNSAFE_componentWillMount() {
         this._isMounted = false;
     }
 
     createPost = async () => {
         try {
+            const userId = await SecureStore.getItemAsync('user_id')
             this._isMounted = true;
-            await fetch(`${apiURL}/post`, {
+            this.setState({
+                isLoading: true,
+            });
+            const response = await fetch(`${apiURL}/post`, {
                 method: 'POST',
                 headers: {
-                    Accept: 'application/json',
                     'Content-Type': 'application/json',
                 },
                 body: JSON.stringify({
                     title: this.state.title,
                     text: this.state.text,
+                    userId: userId
                 }),
             })
-                .then(() => {
-                    this.setState({
-                        newPost: {
-                            title: this.state.title,
-                            text: this.state.text
-                        }
-                    })
-                    console.log("state in add post", this.state)
+
+            const res = await response.json();
+            console.log("res create Post", res);
+            this.setState({
+                isLoading: false
+            })
+            if (response.ok) {
+                //sending new created post data to home page
+                this.props.navigation.state.params.addNewPostToArray({
+                    title: res.title,
+                    text: res.text,
+                    User: res.User,
+                    createdAt: res.createdAt,
+                    id: res.id
                 });
+                this.props.navigation.goBack();
+            }
+            else {
+                alert("Post is not created, please try again!")
+            }
         }
         catch (error) {
             console.error(error);
         }
     }
 
-    componentDidMount() {
-        this.createPost();
-    }
+    /*  componentDidMount() {
+         this.createPost();
+     } */
 
     componentWillUnmount() {
         //this.props.navigation.state.params.getAllPosts();
         //this.props.navigation.state.params.addNewPostToArray(this.state.newPost);
     }
 
-    submitHandler() {
+    submitCreatePost = async () => {
         console.log("combinedFunctionOnSubmit");
-        const { navigation } = this.props;
+        const userId = await SecureStore.getItemAsync('user_id')
         this.createPost();
-        //  navigation.state.params.getAllPosts();
-        navigation.state.params.addNewPostToArray({ title: this.state.title, text: this.state.text });
-        navigation.goBack();
-        //navigation.state.params.onNavigateBack();
     }
-
-    /*   changeHandlerTitle = (title, text) => {
-          console.log("title text", text);
-          if (text.trim().length === 0 && title.trim().length === 0) {
-              this.setState({
-                  titleIsValid: false
-              })
-          } else {
-              this.setState({
-                  titleIsValid: true
-              })
-          }
-          this.setState({
-              title: title,
-              text: text
-          })
-      }
-  
-      submitHandler = () => {
-          const { navigation } = this.props;
-          if (!this.state.titleIsValid) {
-              alert('Please enter a title and description', [
-                  { text: 'Okay' }
-              ]);
-              return;
-          } else {
-              this.createPost();
-              navigation.state.params.addNewPostToArray({ title: this.state.title, text: this.state.text });
-              navigation.goBack();
-          }
-      } */
 
 
     render() {
-        return (
-            <ScrollView>
-                <View style={styles.container}>
-                    <TextInput
-                        placeholder={str.POST_TITLE}
-                        placeholderTextColor="lightgrey"
-                        value={this.state.title}
-                        onChangeText={(title) => this.setState({ title })}
-                        //onChangeText={this.changeHandlerTitle.bind(this, 'title')}
-                        style={styles.input}
-                        autoCorrect
-                        returnKeyType="next"
-                    />
-                    <TextInput
-                        placeholder={str.POST_DETAILS}
-                        placeholderTextColor="lightgrey"
-                        value={this.state.text}
-                        onChangeText={(text) => this.setState({ text })}
-                        //onChangeText={this.changeHandlerTitle.bind(this, 'text')}
-                        style={styles.input}
-                        autoCorrect
-                    />
-
-                    <TouchableOpacity
-                        style={styles.submitButton}
-                        onPress={() => this.submitHandler()}
-                        disabled={!this.state.text || !this.state.title}
-                    >
-                        <Text style={styles.submitButtonText}> Submit </Text>
-                    </TouchableOpacity>
+        if (this.state.isLoading) {
+            return (
+                <View style={styles.centered} >
+                    <ActivityIndicator animating={true} size="large" />
                 </View>
-            </ScrollView>
+            )
+        }
+        return (
+            <View style={styles.container}>
+                <TextInput
+                    placeholder={str.POST_TITLE}
+                    placeholderTextColor="lightgrey"
+                    value={this.state.title}
+                    onChangeText={(title) => this.setState({ title })}
+                    style={styles.input}
+                    autoCorrect
+                    returnKeyType="next"
+                    multiline={true}
+                />
+                <TextInput
+                    placeholder={str.POST_DETAILS}
+                    placeholderTextColor="lightgrey"
+                    value={this.state.text}
+                    onChangeText={(text) => this.setState({ text })}
+                    style={styles.input}
+                    autoCorrect
+                    multiline={true}
+                    spellCheck={true}
+                />
 
+                <TouchableOpacity
+                    style={styles.button}
+                    onPress={() => this.createPost()}
+                    disabled={!this.state.text || !this.state.title}
+                >
+                    <Text style={styles.buttonText}> Submit </Text>
+                </TouchableOpacity>
+            </View>
         )
     }
 }
+AddPostScreen.navigationOptions = navData => {
+    return {
+        headerTitle: 'Create Post',
+        headerStyle: {
+            backgroundColor: Platform.OS === 'android' ? Color.TEAL : ''
+        },
+        headerTintColor: Platform.OS === 'android' ? Color.WHITE : Color.TEAL,
+    };
+};
 
 const styles = StyleSheet.create({
     container: {
         flex: 1,
+        marginTop: 20
+    },
+    centered: {
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center'
     },
     input: {
         height: 40,
         marginBottom: 20,
+        width: "90%",
+        alignSelf: 'center',
         borderColor: "rgba(255,255,255,0.7)",
-        color: color.BLACK,
+        color: Color.BLACK,
         paddingHorizontal: 10,
         borderBottomWidth: 1,
-        borderBottomColor: color.LIGHTGRAY
+        borderBottomColor: Color.LIGHTGRAY
     },
-    submitButton: {
-        backgroundColor: '#fc454e',
-        padding: 10,
-        margin: 15,
-        height: 40,
+    button: {
+        width: "90%",
+        alignSelf: "center",
+        paddingVertical: 12,
+        borderRadius: 4,
+        borderColor: "rgba(255,255,255,0.7)",
+        backgroundColor: Color.TEAL,
     },
-    submitButtonText: {
-        color: 'white'
-    }
+    buttonText: {
+        textAlign: 'center',
+        fontWeight: '700',
+        color: Color.WHITE
+    },
 
 })
 
