@@ -1,50 +1,66 @@
 import React, { Component } from 'react';
 import { Ionicons } from '@expo/vector-icons';
 import { View, TouchableOpacity, StyleSheet, Text } from 'react-native';
-import { Card, Divider } from 'react-native-elements'
-
-import сolor from '../constants/Colors';
+import { Card, Divider, Avatar } from 'react-native-elements'
+import Moment from 'react-moment';
+import Color from '../constants/Colors';
 
 class PostCard extends Component {
     constructor(props) {
         super(props);
         this.state = {
-           // data: props.posts,
             visibleModal: false, // for modal
         };
     }
 
-
+    navigateToPostDetail() {
+        this.props.navigation.navigate({
+            routeName: 'PostDetails',
+            params: {
+                postId: this.props.id,
+                postTitle: this.props.title,
+                postText: this.props.text,
+                createdAt: this.props.date,
+                postUserName: this.props.user
+            }
+        })
+    }
 
     render() {
+        const timeAgo = this.props.date
         return (
-            <Card style={styles.container}>
-                <View style={styles.header}>
-                    <Ionicons style={styles.userIcon} name='md-aperture' size={26} />
-                    <Text>Tomas</Text>
-                    <View style={styles.modalButton}>
-                        <TouchableOpacity
-                            // in parameter send props about each card to delete 
-                            onPress={() => this.props.openModal(this.props.id)}
-                            hitSlop={{ top: 10, bottom: 10, left: 10, right: 10 }}>
-                            <Ionicons name='md-more' size={22} />
-                        </TouchableOpacity>
+            <TouchableOpacity onPress={() => this.navigateToPostDetail()}>
+                <Card>
+                    <View style={styles.header}>
+                        <Avatar size='medium' rounded />
+                        <Text style={styles.titleText}>{this.props.title}</Text>
+                        <View style={styles.modalButton}>
+                            <TouchableOpacity
+                                onPress={() => this.props.openModal(this.props.id)}  // in parameter send props about each card to delete 
+                                hitSlop={{ top: 50, bottom: 50, left: 20, right: 50 }}>
+                                <Ionicons name='md-more' size={22} />
+                            </TouchableOpacity>
+                        </View>
                     </View>
-                </View>
-                <View>
-                    <Text>{this.props.date}</Text>
-                    <Text style={styles.titleText}>{this.props.title}</Text>
-                    <Text style={styles.mainText}>{this.props.text}</Text>
-                </View>
-                <View>
+
+                    <View style={styles.header}>
+                        <Text style={styles.username} >by {this.props.user}</Text>
+                        <Moment style={styles.moment} element={Text} fromNow>{timeAgo}</Moment>
+                    </View>
                     <Divider style={styles.divider} />
-                    <TouchableOpacity style={styles.commentContainer} onPress={() => alert("Add comment later")}>
-                        <Ionicons style={styles.commentIcon} name='md-albums' style={{ color: '#ED4A6A' }} size={20} />
+
+                    <View>
+                        <Text style={styles.mainText}>{this.props.text}</Text>
+                    </View>
+                    <Divider style={styles.divider} />
+
+                    {/* Handle Comment to a Post  */}
+                    <TouchableOpacity style={styles.commentContainer} onPress={() => this.navigateToPostDetail()}>
+                        <Ionicons style={styles.commentIcon} name='md-albums' style={{ color: Color.PASTELRED }} size={20} />
                         <Text style={styles.commentText}>Comment</Text>
                     </TouchableOpacity>
-                </View>
-            </Card>
-
+                </Card>
+            </TouchableOpacity>
         );
     }
 }
@@ -59,41 +75,36 @@ const styles = StyleSheet.create({
         flexDirection: 'row',
         alignItems: 'center',
     },
+    username: {
+        marginLeft: 70,
+        marginRight: 15,
+        color: 'grey'
+    },
     modalButton: {
-        marginLeft: 200,
-        //justifyContent: 'flex-end',
-        // alignContent: 'flex-end',
-        alignSelf: 'flex-end'
-    },
-
-    contentCard: {
-        //flexDirection: 'row',
-        //flexDirection: 'column-reverse'
-        backgroundColor: 'green',
-        height: 40
-    },
-
-    card: {
-        backgroundColor: 'red'
+        position: 'absolute',
+        right: 5
     },
 
     userIcon: {
         marginRight: 10
     },
     titleText: {
-        fontSize: 20,
-        paddingBottom: 0,
-        paddingTop: 0
+        fontSize: 24,
+        paddingLeft: 20,
+    },
+    moment: {
+        color: 'grey'
     },
     mainText: {
         paddingBottom: 15
     },
     divider: {
+        marginBottom: 15,
+        marginTop: 15,
     },
     commentContainer: {
         flexDirection: 'row',
-        alignItems: 'center',
-        textAlign: 'justify',
+        justifyContent: 'flex-end',
     },
     commentText: {
         marginLeft: 10,
@@ -113,3 +124,7 @@ export default PostCard;
 
 
 // https://docs.nativebase.io/Components.html#card-list-headref
+
+
+//moment to manage timeago
+//https://github.com/headzoo/react-moment
