@@ -1,7 +1,7 @@
 import React from 'react';
 import { Ionicons } from '@expo/vector-icons';
 
-import color from '../constants/Colors';
+import Color from '../constants/Colors';
 
 export default function TabBarIcon(props) {
   return (
@@ -9,7 +9,7 @@ export default function TabBarIcon(props) {
       name={props.name}
       size={26}
       style={{ marginBottom: -3 }}
-      color={props.focused ? color.tabIconSelected : color.tabIconDefault}
+      color={props.focused ? Color.TEAL : Color.tabIconDefault}
     />
   );
 }
