@@ -1,4 +1,4 @@
-import { React } from "react";
+/* import { React } from "react";
 import { createAppContainer, createSwitchNavigator } from "react-navigation";
 
 import MainTabNavigator from './MainTabNavigator';
@@ -8,4 +8,4 @@ export default createAppContainer(
     createSwitchNavigator({
         Main: MainTabNavigator,
     })
-);
+); */
